Tests the Tools custom editor functionality
This Play Mode test simulates what happens when you press the Enable and Disable button on the custom inspector for the Tools script.

There's no way to drive the editor from a test so this extracts the functionality and the test executes the methods as if the custom inspector button was pressed.